Bio system working, next is remove extras

+++ src/Controller/BiosController.php
@@ -2,6 +2,8 @@
 namespace App\Controller;
 
 use App\Controller\AppController;
+use Cake\Core\Configure;
+use Cake\I18n\Time;
 
 /**
  * Bios Controller
@@ -18,10 +20,16 @@ class BiosController extends AppController
      *
      * @return \Cake\Http\Response|void
      */
+
     public function index()
     {
         $this->paginate = [
-            'contain' => ['Users', 'Purposes']
+            'contain' => ['Users', 'Purposes'],
+            'sortWhitelist' => [
+                'Bios.id',
+                'Users.last',
+                'Bios.purpose_id'
+            ],
         ];
 
         if ( $this->Auth->user('is_admin')) {
@@ -30,6 +38,11 @@ class BiosController extends AppController
             $bios = $this->paginate($this->Bios->find('all')
                 ->where(['user_id' => $this->Auth->user('id')]));
         }
+
+        $this->set('crumby', [
+            ["/", __("Dashboard")],
+            [null, __("Bios")]
+        ]);
 
         $this->set('isAdmin', $this->Auth->user('is_admin'));
         $this->set('bios', $bios);
@@ -45,9 +58,19 @@ class BiosController extends AppController
      */
     public function view($id = null)
     {
+        
+        
         $bio = $this->Bios->get($id, [
             'contain' => ['Users', 'Purposes']
         ]);
+
+        $this->set('crumby', [
+            ["/", __("Dashboard")],
+            ["/bios/", __("Bios")],
+            [null, $bio->user->print_name . " - " . $bio->purpose->name ]
+        ]);
+
+        $this->set('tz', Configure::read('ServerTimeZoneFix'));
 
         if ( $this->Auth->user('is_admin') || $bio->user_id == $this->Auth->user('id') ) {
             $this->set('bio', $bio);
@@ -67,15 +90,38 @@ class BiosController extends AppController
         $bio = $this->Bios->newEntity();
         if ($this->request->is('post')) {
             $bio = $this->Bios->patchEntity($bio, $this->request->getData());
-            if ($this->Bios->save($bio)) {
-                $this->Flash->success(__('The bio has been saved.'));
+            if ( $this->Auth->user('is_admin') || $bio->user_id == $this->Auth->user('id') ) {
+                if ($this->Bios->save($bio)) {
+                    $this->Flash->success(__('The bio has been saved.'));
 
-                return $this->redirect(['action' => 'index']);
+                    return $this->redirect(['action' => 'index']);
+                }
+                $this->Flash->error(__('The bio could not be saved. Please, try again.'));
+            } else {
+                $this->Flash->error(__('You may only add your own bios.'));
+                return $this->redirect("/bios");
             }
-            $this->Flash->error(__('The bio could not be saved. Please, try again.'));
         }
-        $users = $this->Bios->Users->find('list', ['limit' => 200]);
-        $purposes = $this->Bios->Purposes->find('list', ['limit' => 200]);
+
+        $userWhere = [
+            'is_active' => 1,
+            'is_verified' => 1,
+        ];
+
+        if ( ! $this->Auth->user('is_admin') ) {
+            $userWhere['id'] = $this->Auth->user('id');
+        }
+
+        $this->set('crumby', [
+            ["/", __("Dashboard")],
+            ["/bios/", __("Bios")],
+            [null, __("Add Bio")]
+        ]);
+
+        $users = $this->Bios->Users->find('list')
+            ->where($userWhere)
+            ->order(['Users.last' => 'ASC']);
+        $purposes = $this->Bios->Purposes->find('list', ['limit' => 200])->where(['is_active' => 1 ]);
         $this->set(compact('bio', 'users', 'purposes'));
     }
 
@@ -93,15 +139,27 @@ class BiosController extends AppController
         ]);
         if ($this->request->is(['patch', 'post', 'put'])) {
             $bio = $this->Bios->patchEntity($bio, $this->request->getData());
-            if ($this->Bios->save($bio)) {
-                $this->Flash->success(__('The bio has been saved.'));
+            if ( $this->Auth->user('is_admin') || $bio->user_id == $this->Auth->user('id') ) {
+                if ($this->Bios->save($bio)) {
+                    $this->Flash->success(__('The bio has been saved.'));
 
-                return $this->redirect(['action' => 'index']);
+                    return $this->redirect(['action' => 'index']);
+                }
+                $this->Flash->error(__('The bio could not be saved. Please, try again.'));
+            } else {
+                $this->Flash->error(__('You may only edit your own bios.'));
+                return $this->redirect("/bios");
             }
-            $this->Flash->error(__('The bio could not be saved. Please, try again.'));
         }
-        $users = $this->Bios->Users->find('list', ['limit' => 200]);
-        $purposes = $this->Bios->Purposes->find('list', ['limit' => 200]);
+
+        $this->set('crumby', [
+            ["/", __("Dashboard")],
+            ["/bios/", __("Bios")],
+            [null, __("Edit Bio")]
+        ]);
+
+        $users = $this->Bios->Users->find('list', ['limit' => 200])->where(['id' => $bio->user_id]);
+        $purposes = $this->Bios->Purposes->find('list', ['limit' => 200])->where(['is_active' => 1 ]);
         $this->set(compact('bio', 'users', 'purposes'));
     }
 
@@ -116,10 +174,16 @@ class BiosController extends AppController
     {
         $this->request->allowMethod(['post', 'delete']);
         $bio = $this->Bios->get($id);
-        if ($this->Bios->delete($bio)) {
-            $this->Flash->success(__('The bio has been deleted.'));
+        if ( $this->Auth->user('is_admin') || $bio->user_id == $this->Auth->user('id') ) {
+
+            if ($this->Bios->delete($bio)) {
+                $this->Flash->success(__('The bio has been deleted.'));
+            } else {
+                $this->Flash->error(__('The bio could not be deleted. Please, try again.'));
+            }
         } else {
-            $this->Flash->error(__('The bio could not be deleted. Please, try again.'));
+            $this->Flash->error(__('You may only delete your own bios.'));
+            return $this->redirect("/bios");
         }
 
         return $this->redirect(['action' => 'index']);
